% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamsel-internal.R
\name{gamsel-internal}
\alias{gamsel-internal}
\alias{basis.subset}
\alias{df.inv}
\alias{fracdev}
\alias{getmin}
\alias{norm2}
\alias{poly}
\alias{summarynz}
\alias{summaryplot}
\alias{error.bars}
\title{Internal gamsel functions}
\description{
These are not intended for use by users.
}
\author{
Trevor Hastie
}
\keyword{internal}
