\name{gamsel-internal}
\title{Internal gamsel functions}
\alias{basis.subset}
\alias{df.inv}
\alias{fracdev}
\alias{gendata}
\alias{getmin}
\alias{norm2}
\alias{poly}
\alias{getmin}
\alias{summarynz}
\alias{summaryplot}
\alias{error.bars}
\description{Internal gamsel functions}
\usage{
basis.subset(basis.object,subset,...)
}
\author{Trevor Hastie}
\details{These are not intended for use by users.
}
\keyword{internal}
