% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamsel-package.R
\docType{package}
\name{gamsel-package}
\alias{gamsel-package}
\title{Fit Regularization Path for Generalized Additive Models}
\description{
Using overlap grouped lasso penalties, gamsel selects whether a
term in a gam is nonzero, linear, or a non-linear spline (up to a
specified max df per variable). It fits the entire regularization
path on a grid of values for the overall penalty lambda, both for
gaussian and binomial families. Key functions are \link{gamsel},
\link{predict.gamsel}, \link{plot.gamsel}, \link{print.gamsel}, \link{summary.gamsel},
\link{cv.gamsel}, \link{plot.cv.gamsel}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arxiv.org/abs/1506.03850}
}

}
\author{
Alexandra Chouldechova, Trevor Hastie Maintainer: Trevor Hastie
\email{hastie@stanford.edu}
}
\keyword{additive}
\keyword{generalized}
\keyword{models}
\keyword{package}
\keyword{regression}
